Mountain Max Height Calculator
This tool generates a window that allows a user to enter mass and radius values. Using the equation for hydrostatic equilibrium it then calculates the maximum height of a given mountain on such a planet of the entered values. The system is simplified and assumptions like constant density and a specific bulk modulus are made.